==> Ventas.df.kaggle/1.-ETL.ipynb <==
***1.Introcuccion de este proyecto*** <br><br>
Una hoja de cálculo que tenga, como mínimo, lo siguiente:<br><br>
Fuente de datos pre-procesada (datos limpios)<br><br>
Una pestaña de reporte con tablas y de las métricas principales del negocio<br><br>
Una pestaña con un análisis por cohorte general y también uno para clientes UK y clientes no-UK<br><br>
Una segmentación usando la metodología RFM<br><br>
Todos los datos, tablas y visualizaciones adicionales que te sumen en tu análisis<br><br>




Fuente de datos pre-procesada (datos limpios)

#importamos libreria
import pandas as pd

#cargamos datos mediante libreria 
df = pd.read_csv(r'C:\Users\ga8ri\vscode\ETL\Ventas.df.kaggle\ventas-por-factura.csv')
df.head()

#calculamos numero de valores nulos 
nulos_por_columna = df.isnull().sum()
nulos_por_columna

#extraemos informacion del nombre de las columnas 
df.columns
#obtenemos forrmato de los datos 
df.info()

#reemplazamos la , por un . para la posterior conversion a FLOAT
if df['Monto'].dtype == 'object':
    df['Monto'] = df['Monto'].str.replace(',', '.')
    df['Monto'] = df['Monto'].astype(float)
df


# Crear la columna "Tipo de factura" 
#NC nota d credito
#F Factura
df['Tipo de factura'] = df['N° de factura'].apply(lambda x: 'ND' if 'C' in str(x) else 'F')

# Mostrar el DataFrame con la nueva columna
df


#reordenar datos acuerdo al N de factura
df.sort_values(by='N° de factura')
df

#longitud del dataFrame 
longitud_original = len(df)

# Eliminar los duplicados 
df.drop_duplicates()

# Calcular cuántos datos fueron eliminados
datos_eliminados = longitud_original - len(df)

print("Se eliminaron", datos_eliminados, "datos al eliminar los duplicados.")

***DF para anlisis general listo***

--------------------------------------------------------------------------------------------------------------------------

***DF CLIENTES***

#preparamos la data para el analisis RFM 
df_clientes = df


#Eliminamos filas donde no este el ID CLIENTE 
df_clientes = df_clientes.dropna(subset=['ID Cliente'])

#Se transforma de int a str 
df_clientes['ID Cliente'] = df_clientes['ID Cliente'].astype(str)

df_clientes


#pasamos fecha de factura a datatime 
df_clientes['Fecha de factura'] = pd.to_datetime(df['Fecha de factura'])
#extraemos año,mes,dia y hora en nuevas columnas 
df_clientes['Año'] = df_clientes['Fecha de factura'].dt.year
df_clientes['Mes'] = df_clientes['Fecha de factura'].dt.month
df_clientes['Día'] = df_clientes['Fecha de factura'].dt.day
df_clientes['Hora'] = df_clientes['Fecha de factura'].dt.hour

df_clientes


# Crear un diccionario con los nombres de los días de la semana en español
dias_semana = {
    0: 'lunes',
    1: 'martes',
    2: 'miércoles',
    3: 'jueves',
    4: 'viernes',
    5: 'sábado',
    6: 'domingo'
}

# Crear la columna 'Día de la semana' que contiene el nombre del día en español
df_clientes['Día de la semana'] = df_clientes['Fecha de factura'].dt.dayofweek.map(dias_semana)

# Mostrar el DataFrame con la nueva columna 'Día de la semana'
df_clientes


# Obtener la frecuencia de cada valor en la columna "N° de factura"
frecuencia = df['N° de factura'].value_counts()

# Filtrar los valores que se repiten
valores_repetidos = frecuencia[frecuencia > 1].index.tolist()

# Mostrar la lista de valores que se repiten
print(len(valores_repetidos))

filas_repetidas = df_clientes[df_clientes['N° de factura'].isin(valores_repetidos)]

# Mostrar las filas seleccionadas
filas_repetidas

# Agregar una nueva columna 'NumeroOrden' que enumere las facturas
df_clientes['NumeroOrden'] = df_clientes.groupby(['N° de factura']).ngroup() + 1

# Mostrar el DataFrame con la nueva columna 'NumeroOrden'
df_clientes


# Guardar el DataFrame como un archivo CSV
df_clientes.to_csv('clientes.csv', index=False)

----------------------------------------------------------

Seguimos con nuetro df para analisis general 

df['Fecha de factura'] = pd.to_datetime(df['Fecha de factura'])

df['Año'] = df['Fecha de factura'].dt.year
df['Mes'] = df['Fecha de factura'].dt.month
df['Día'] = df['Fecha de factura'].dt.day
df['Hora'] = df['Fecha de factura'].dt.hour

# Crear un diccionario con los nombres de los días de la semana en español
dias_semana = {
    0: 'lunes',
    1: 'martes',
    2: 'miércoles',
    3: 'jueves',
    4: 'viernes',
    5: 'sábado',
    6: 'domingo'
}

# Crear la columna 'Día de la semana' que contiene el nombre del día en español
df['Día de la semana'] = df['Fecha de factura'].dt.dayofweek.map(dias_semana)

# Mostrar el DataFrame con la nueva columna 'Día de la semana'
df


df.to_csv('data_limpia.csv', index=False)
