==> Ventas.df.kaggle/2.-datalimpia.ipynb <==
import pandas as pd

df = pd.read_csv(r'C:\Users\ga8ri\vscode\ETL\Ventas.df.kaggle\data_limpia.csv')
df

df.columns

        1.-¿Cuánto es lo que venden al mes?
        2.-¿cuántas ventas tienen en cada mes?
        3.-¿cuál es el mes que más venden en dolares?
        4.-¿cómo se desglosa esta información según los países?

**1.-¿Cuánto es lo que venden en promedio al mes?**


# Diccionario de meses
meses_espanol = {
    1: 'enero',
    2: 'febrero',
    3: 'marzo',
    4: 'abril',
    5: 'mayo',
    6: 'junio',
    7: 'julio',
    8: 'agosto',
    9: 'septiembre',
    10: 'octubre',
    11: 'noviembre',
    12: 'diciembre'
}
#DataFrame para excluir las filas con "ND" en la columna Tipo de factura
df_sin_nd = df[df['Tipo de factura'] != 'ND']

# Calcular el promedio de ventas por mes
suma_ventas_mes = df_sin_nd.groupby('Mes')['Monto'].sum().round(2)

# Crear un DataFrame con los resultados y agregar el nombre del mes en español
df_promedio_ventas_mes = pd.DataFrame({
    'Mes': suma_ventas_mes.index.map(meses_espanol),
    'Suma de ventas': suma_ventas_mes.values
})

# Imprimir la tabla
print(df_promedio_ventas_mes)

# Calcular el promedio total de todos los meses
promedio_total = df_sin_nd['Monto'].sum().round(2)
print("La suma total de ventas es:", promedio_total)


 2.-¿cuántas ventas tienen en cada mes?

# Agrupar las ventas por mes y contar el número de ventas en cada mes
ventas_por_mes = df.groupby('Mes').size().reset_index(name='Ventas')

#Columna del nombre del mes
ventas_por_mes['Nombre del mes'] = ventas_por_mes['Mes'].map(meses_espanol)

ventas_por_mes


3.-¿cuál es el mes que más venden en dolares?


mes_con_mas_ventas = df_promedio_ventas_mes.loc[df_promedio_ventas_mes['Suma de ventas'].idxmax()]

print("El mes en el que se vendió más en dólares fue:", mes_con_mas_ventas['Mes'], "con un total de", mes_con_mas_ventas['Suma de ventas'], "dólares.")


#Suma de ventas por país
suma_ventas_por_pais = df_sin_nd.groupby('País')['Monto'].sum().round(2)

# Seleccionar los 10 países con la mayor suma de ventas
top_10_ventas_por_pais = suma_ventas_por_pais.nlargest(10)

# Crear un Df
df_top_10_ventas_por_pais = pd.DataFrame({
    'País': top_10_ventas_por_pais.index,
    'Suma de ventas': top_10_ventas_por_pais.values
})

# Imprimir la tabla
df_top_10_ventas_por_pais


# Calcular la suma de ventas por país y mes
ventas_por_pais_y_mes = df_sin_nd.groupby(['País', 'Mes'])['Monto'].sum().round(2)

# Seleccionar los 10 países con la mayor suma de ventas
top_10_paises = suma_ventas_por_pais.nlargest(10).index

# Filtrar el DataFrame original para incluir solo los 10 países principales
df_top_10_ventas_por_pais_y_mes = ventas_por_pais_y_mes.loc[top_10_paises]

# Restablecer el índice
df_top_10_ventas_por_pais_y_mes = df_top_10_ventas_por_pais_y_mes.reset_index()

# Realizar la operación de pivote para tener los meses como columnas
df_top_10_ventas_por_pais_y_mes = df_top_10_ventas_por_pais_y_mes.pivot(index='País', columns='Mes', values='Monto')

# Imprimir el DataFrame
df_top_10_ventas_por_pais_y_mes
