==> Ventas.df.kaggle/3.-clientes.ipynb <==
***En este cuaderno se responden las siguientes preguntas segun el problema planteado***<br><br>


1.-¿Cuánto es lo que venden en promedio al mes a los clientes?,<br> 
2.-¿cuántas ventas tienen en cada mes de los clientes?, <br>
3.-¿cuál es el mes que más venden en dolares a los clientes?, <br>
4.-¿quiénes son sus clientes más importantes?, <br>
5.-¿cuál es el monto promedio que gastan sus clientes?, <br>
6.-¿qué porcentaje de sus clientes han vuelto a comprarles?, <br>
7.-¿cómo se desglosa esta información según los países en donde venden a los clientes?”<br>


#Cargamos librerias 
import pandas as pd

#cargamos data clientes
df = pd.read_csv(r'C:\Users\ga8ri\vscode\ETL\Ventas.df.kaggle\clientes.csv')
df

# Seleccionar las filas con el mismo valor absoluto en 'Cantidad' y 'Monto' y el mismo ID de cliente
mismo_valor_absoluto_mismo_cliente = df[(abs(df['Cantidad']) == abs(df['Monto'])) & (df['ID Cliente'].duplicated(keep=False))]

mismo_valor_absoluto_mismo_cliente


#obtenemos la informaicon correspondiente 
df.info()
df.columns

***1-Monto promedio por mes***


# Calcular la suma de montos vendidos por mes
ventas_por_mes = df.groupby('Mes')['Monto'].sum()

# Calcular el promedio de ventas por mes
promedio_ventas_por_mes = round(ventas_por_mes.mean(),2)

# Mostrar el promedio de ventas por mes
print("El promedio de ventas por mes es:", promedio_ventas_por_mes,'de dolares')


***2.-¿cuántas ventas tienen en cada mes de los clientes?***

# Contar el número de ventas en cada mes y redondear a 2 decimales
ventas_por_mes = df.groupby('Mes').size().round(2)

# Ordenar en orden descendente
ventas_por_mes = ventas_por_mes.sort_values(ascending=False)

meses_espanol = {
    1: 'enero', 2: 'febrero', 3: 'marzo', 4: 'abril', 5: 'mayo', 6: 'junio',
    7: 'julio', 8: 'agosto', 9: 'septiembre', 10: 'octubre', 11: 'noviembre', 12: 'diciembre'
}


tabla_ventas = ventas_por_mes.to_frame(name='cantidad de ventas')
tabla_ventas['Mes.'] = tabla_ventas.index.map(meses_espanol)

# Mostrar el número de ventas por mes ordenado en orden descendente
print(tabla_ventas)


***3.-¿cuál es el mes que más venden en dolares a los clientes?***

# Encontrar el mes con más ventas en dolares
mes_mas_ventas = ventas_por_mes.idxmax()

# Encontrar el monto total vendido en el mes con más ventas
monto_mes_mas_ventas = df[df['Mes'] == mes_mas_ventas]['Monto'].sum()

# Mostrar el mes con más ventas y el monto total vendido en ese mes
print("El mes con más ventas es:", mes_mas_ventas)
print("El monto total vendido en ese mes es:", round(monto_mes_mas_ventas,2))


***5.-¿cuál es el monto promedio que gastan sus clientes?,***

# Calcular el monto promedio que gastan los clientes
monto_promedio_por_cliente = df.groupby('ID Cliente')['Monto'].mean()

# Calcular el promedio del monto promedio por cliente
promedio_total = monto_promedio_por_cliente.mean().round(2)

# Mostrar el monto promedio que gastan los clientes y el promedio total
print("El monto promedio que gastan los clientes es:", promedio_total)


***6.-¿qué porcentaje de sus clientes han vuelto a comprarles?,***

# Contar el número de órdenes únicas por cliente
ordenes_por_cliente = df.groupby('ID Cliente')['NumeroOrden'].nunique()

# Contar el número de clientes que han hecho más de una orden
clientes_recurrentes = (ordenes_por_cliente > 1).sum()

 
porcentaje_clientes_recurrentes = ((clientes_recurrentes / len(ordenes_por_cliente)) * 100).round(2)

# Mostrar el porcentaje de clientes recurrentes
print("El numero de clientes que han vuelto a comprarles es:", clientes_recurrentes)
print("El porcentaje de clientes que han vuelto a comprarles es:", porcentaje_clientes_recurrentes, "%")


***7.-¿cómo se desglosa esta información según los países en donde venden a los clientes?***

# número total de clientes
cantidad_total_clientes = df['ID Cliente'].nunique()

# Cantidad total de ventas por país
cantidad_por_pais = df.groupby('País')['NumeroOrden'].nunique()

# Calcular el monto total de ventas por país
monto_por_pais = df.groupby('País')['Monto'].sum()

# Crear un DataFrame con los resultados
tabla_resultados = pd.DataFrame({
    'Cantidad total de clientes': [cantidad_total_clientes],
    'Cantidad total de ventas': 
    [cantidad_por_pais.sum()],
    'Monto total de ventas': 
    [monto_por_pais.sum()]
})

# Agregar los resultados por país al DataFrame
tabla_resultados_por_pais = pd.DataFrame({
    'Cantidad de clientes': df.groupby('País')['ID Cliente'].nunique(),
    'Cantidad de ventas': cantidad_por_pais,
    'Monto de ventas': monto_por_pais
})



tabla_resultados_por_pais = tabla_resultados_por_pais.sort_values(by=['Monto de ventas', 'Cantidad de clientes', 'Cantidad de ventas'], ascending=False)
tabla_resultados_por_pais


***4.-¿quiénes son sus clientes más importantes?   RFM***

Recencia o Actualidad, Frecuencia y Valor Monetario

# Crear df_segmentacion excluyendo las filas con tipo de factura 'ND'
df_segmentacion = df.loc[df['Tipo de factura'] != 'ND', ['N° de factura', 'Fecha de factura', 'ID Cliente', 'Monto']]

# Imprimir el DataFrame df_segmentacion
df_segmentacion



# Obtener la última fecha de compra por ID Cliente
ultima_compra = df_segmentacion.groupby('ID Cliente')['Fecha de factura'].max().reset_index()

# Calcular la fecha máxima de la tabla
fecha_maxima = df_segmentacion['Fecha de factura'].max()

# Calcular los días transcurridos desde la última compra
ultima_compra['Días transcurridos'] = (fecha_maxima - ultima_compra['Fecha de factura']).dt.days

# Obtener el monto total por ID Cliente
monto = df_segmentacion.groupby('ID Cliente')['Monto'].sum().reset_index()

# Renombrar las columnas
ultima_compra = ultima_compra.rename(columns={'Fecha de factura': 'Última compra'})
monto = monto.rename(columns={'Monto': 'Monto total'})

# Contar el número de compras por ID Cliente
num_compras = df_segmentacion.groupby('ID Cliente').size().reset_index(name='Num compras')



# Combinar los DataFrames
df_resultado = pd.merge(ultima_compra, monto, on='ID Cliente')
# Combinar los DataFrames
df_resultado = pd.merge(df_resultado, num_compras, on='ID Cliente')

# Imprimir el resultado
df_resultado


# Calculamos los límites de los quintiles
quintil_1 = df_resultado['Días transcurridos'].min()
quintil_2 = df_resultado['Días transcurridos'].quantile(0.25)
quintil_3 = df_resultado['Días transcurridos'].quantile(0.5)
quintil_4 = df_resultado['Días transcurridos'].quantile(0.75)
quintil_5 = df_resultado['Días transcurridos'].max()

# Función para asignar puntajes según el quintil
def asignar_puntaje(recencia):
    if recencia <= quintil_1:
        return 5
    elif recencia <= quintil_2:
        return 4
    elif recencia <= quintil_3:
        return 3
    elif recencia <= quintil_4:
        return 2
    else:
        return 1

# Calcular el puntaje de recencia para cada cliente
df_resultado['Puntaje recencia'] = df_resultado['Días transcurridos'].apply(asignar_puntaje)

df_resultado


# Calculamos los límites de los quintiles para la columna "Num compras"
quintil_1 = df_resultado['Num compras'].min()
quintil_2 = df_resultado['Num compras'].quantile(0.25)
quintil_3 = df_resultado['Num compras'].quantile(0.5)
quintil_4 = df_resultado['Num compras'].quantile(0.75)
quintil_5 = df_resultado['Num compras'].max()

# Función para asignar puntajes según el quintil
def asignar_puntaje(num_compras):
    if num_compras <= quintil_1:
        return 1
    elif num_compras <= quintil_2:
        return 2
    elif num_compras <= quintil_3:
        return 3
    elif num_compras <= quintil_4:
        return 4
    else:
        return 5

# Calcular el puntaje de num_compras para cada cliente
df_resultado['Puntaje num_compras'] = df_resultado['Num compras'].apply(asignar_puntaje)

df_resultado


# Calculamos los límites de los quintiles para la columna "Monto total"
quintil_1 = df_resultado['Monto total'].min()
quintil_2 = df_resultado['Monto total'].quantile(0.25)
quintil_3 = df_resultado['Monto total'].quantile(0.5)
quintil_4 = df_resultado['Monto total'].quantile(0.75)
quintil_5 = df_resultado['Monto total'].max()

# Función para asignar puntajes según el quintil
def asignar_puntaje(monto):
    if monto <= quintil_1:
        return 1
    elif monto <= quintil_2:
        return 2
    elif monto <= quintil_3:
        return 3
    elif monto <= quintil_4:
        return 4
    else:
        return 5

# Calcular el puntaje de Monto total para cada cliente
df_resultado['Puntaje Monto total'] = df_resultado['Monto total'].apply(asignar_puntaje)

df_resultado


df_resultado['Puntaje RFM'] = df_resultado['Puntaje recencia']* df_resultado['Puntaje num_compras'] * df_resultado['Puntaje Monto total']
df_resultado

# Calculamos el promedio de los puntajes de recencia, número de compras y monto total
df_resultado['Igualdad'] = round((df_resultado['Puntaje recencia'] + df_resultado['Puntaje num_compras'] + df_resultado['Puntaje Monto total']) / 3)
df_resultado


promedio_puntaje_rfm = df_resultado['Puntaje RFM'].mean()
promedio_puntaje_rfm.round()

# Realizamos un conteo de los valores únicos en la columna "igualdad" y los ordenamos por el valor de "igualdad" en orden descendente
conteo_igualdad_ordenado = df_resultado['Igualdad'].value_counts().sort_index(ascending=False)
conteo_igualdad_ordenado


#conteo de los valores únicos en la columna "igualdad"
conteo_igualdad_ordenado = df_resultado['igualdad'].value_counts().sort_index(ascending=False)

# total de observaciones en la columna "igualdad"
total_observaciones = conteo_igualdad_ordenado.sum()

#porcentaje de cada punto
porcentaje_puntos = round((conteo_igualdad_ordenado / total_observaciones) * 100,2)

# Concatenamos el conteo ordenado y el porcentaje en un DataFrame
resultado_final = pd.concat([conteo_igualdad_ordenado, porcentaje_puntos], axis=1)
resultado_final.columns = ['Conteo', 'Porcentaje']

resultado_final


***Los clientes mas importantes son los que correponden a la categoria 5 del RFM ***
